+++ src/components/base/b-router/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2022-??-??)
+
+#### :rocket: New Feature
+
+* Added possibility to disable specific link interception
+
+## v3.30.2 (2022-11-17)
+
+#### :bug: Bug Fix
+
+* Fixed soft transitions with array parameters
+
 ## v3.24.2 (2022-08-19)
 
 #### :house: Internal